底层电机控制完成

+++ include/serial_input.h
@@ -2,5 +2,5 @@
 #include "stdint.h"
 
 
-void read_serial_input(uint8_t* buffer, uint32_t length);
+void read_serial_input(uint8_t* buffer, uint16_t length);
